Grade de aulas corrigida.
-Add aulas duplicadas para dias ou horários diferentes.
-Removendo corretamente todas as aulas e horários de uma determinada
turma, antes de salvar a grade de horários.
-Carregando as definições de grade de aulas, para turmas diferentes

-Alterei a estrutura do POJO de HoraAula e HoraAulaPK
