+++ src/main/java/br/com/sistema/redAmber/basicas/HoraAula.java
@@ -2,32 +2,26 @@ package br.com.sistema.redAmber.basicas;
 
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
-import java.util.Date;
 
-import br.com.sistema.redAmber.basicas.enums.DiasSemana;
 import br.com.sistema.redAmber.basicas.enums.StatusHoraAula;
 
 public class HoraAula {
 
 	private HoraAulaPK id;
 	
-	private DiasSemana dia;
-	
-	private Date horaInicio;
-	
-	private Date horaFim;
+
 	
 	private StatusHoraAula status;
 	
 	@Override
 	public String toString() {
 		
-		if (this.horaInicio != null && this.horaFim != null) {
+		if (this.getId() != null && this.getId().getHoraInicio() != null && this.getId().getHoraFim() != null) {
 			DateFormat df = new SimpleDateFormat("HH:mm");
-			String horaIniStr = df.format(this.getHoraInicio());
-			String horaFimStr = df.format(this.getHoraFim());
+			String horaIniStr = df.format(this.getId().getHoraInicio());
+			String horaFimStr = df.format(this.getId().getHoraFim());
 
-			String retorno = this.dia.toString() + "(Início: " + horaIniStr + " Fim: " + horaFimStr + ")";
+			String retorno = this.getId().getDia().toString() + "(Início: " + horaIniStr + " Fim: " + horaFimStr + ")";
 
 			return retorno;
 		}
@@ -46,21 +40,6 @@ public class HoraAula {
 		this.id = id;
 	}
 
-	public Date getHoraInicio() {
-		return horaInicio;
-	}
-	
-	public void setHoraInicio(Date horaInicio) {
-		this.horaInicio = horaInicio;
-	}
-
-	public Date getHoraFim() {
-		return horaFim;
-	}
-
-	public void setHoraFim(Date horaFim) {
-		this.horaFim = horaFim;
-	}
 
 	public StatusHoraAula getStatus() {
 		return status;
@@ -70,11 +49,4 @@ public class HoraAula {
 		this.status = status;
 	}
 
-	public DiasSemana getDia() {
-		return dia;
-	}
-
-	public void setDia(DiasSemana dia) {
-		this.dia = dia;
-	}
 }
